backend/requirements/base.txt
Django==3.0.6
django-filter==2.2.0
djangorestframework==3.11.0
feedparser==5.2.1
drf-yasg==1.17.1
drf-jwt==1.16.1